Fix finalization tests
