+++ packages/universal/shared/src/lifetimes.ts
@@ -53,6 +53,8 @@ if (!lifetime) {
 
     constructor() {
       this.#createRegistry = () => {
+        console.trace();
+
         const Registry = coordination.testing?.registry ?? FinalizationRegistry;
         return new Registry((state) => void state.finalize());
       };
